fix: audit Linus #2 — P0/P1 corrigés, dead code supprimé
P0: RateLimiter.php — fopen() non vérifié → crash si retourne false
P1: DashboardController — (int) cast tronque poids float → (float)
P1: shell.js — setInterval notification jamais clearé (memory leak HTMX)
P1: shell.js — href/icon non échappés dans attributs data-href search
P1: members.js — recherche sans debounce (re-render chaque keystroke)
P1: config/config.php — fichier legacy orphelin supprimé (DB_USERNAME vs DB_USER)
P1: index.html — auth_login.php → auth_login (cohérence routes)
P2: email-templates — data-page="settings" → "admin" (sidebar highlight)

Tests: 2739 pass, 0 fail.

+++ app/Controller/DashboardController.php
@@ -95,9 +95,9 @@ final class DashboardController extends AbstractController {
                 $t = $ballotRepo->tally($currentMotionId, $tenantId);
                 $data['current_motion_votes'] = [
                     'ballots_count' => (int) $t['total_ballots'],
-                    'weight_for' => (int) $t['weight_for'],
-                    'weight_against' => (int) $t['weight_against'],
-                    'weight_abstain' => (int) $t['weight_abstain'],
+                    'weight_for' => (float) $t['weight_for'],
+                    'weight_against' => (float) $t['weight_against'],
+                    'weight_abstain' => (float) $t['weight_abstain'],
                 ];
             }
